mistral_3.1_samll_base/requirements.txt
vllm>=0.4.0
Pillow
requests
torch>=2.0.0
transformers>=4.38.0
accelerate>=0.21.0
mistral-common